camper_nerf_project/data_validation/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
flask-cors==4.0.0 